refactor(sysmon): move sysmon store to src/workspaces/sysmon/
