+++ src/components/sysmon/SysmonWorkspace.tsx
@@ -1,5 +1,5 @@
 import { tokens, Button, Spinner, Tab, TabList } from "@fluentui/react-components";
-import { useSysmonStore, type SysmonWorkspaceTab } from "../../stores/sysmon-store";
+import { useSysmonStore, type SysmonWorkspaceTab } from "../../workspaces/sysmon/sysmon-store";
 import { useAppActions } from "../layout/Toolbar";
 import { SysmonEventTable } from "./SysmonEventTable";
 import { SysmonSummaryView } from "./SysmonSummaryView";
